refactor(stats): SPI-ify StatisticsToadlet
Add a dedicated statistics runtime port and detached page snapshot so the legacy statistics toadlet no longer keeps live Node or NodeClientCore instance dependencies.

Move the legacy statistics-page traversal and most HTML generation into a daemon-side adapter, keep request-context-only placeholder injection in the HTTP layer, and cover the overview/requesters split with focused tests.
